Ensure user and group IDs stay consistent
After the last reboot it happened that vaultwarden's UID changed, and it
could not access its files anymore.

`tmpfiles.d` only fixes the owner of the FOLDER, but not the content
therein. In theory it would be possible to fix this via `tmpfiles.d` as
well (via `Z` directives), but that could end up in an endless hunt for
broken UID/GID mappings.

So instead back up `/var/lib/nixos`, which contains the UID and GID
mappings.

Signed-off-by: Manuel Hutter <manuel@hutter.io>