PostgreSQL: Try to prevent "Permission denied"
Every so often, writes (and reads?) to PostgreSQL start failing, with
lots of "Permission denied" error messages in journald.

Initially I suspected the issue to occur when booting the server, but
further investigation showed that the issue only occurs after deploying
a new NixOS configuration.

I initially suspected the mount unit to be the culprit, but if I checked
the right logs, it's the "systemd-tmpfiles-resetup.service" unit instead
that messes with PostgreSQL.

I'll leave this active for a bit, time will tell.

Signed-off-by: Manuel Hutter <manuel@hutter.io>